backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.40.1
torch==2.3.0
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-docx==1.1.0
lxml==5.2.1
sentencepiece==0.2.0
accelerate==0.30.1
